# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Atlantis Authentication API
  description: |
    Production-ready authentication and authorization service built with Rust and REST API architecture.

    This service provides secure user registration, JWT authentication, and password reset functionality.

    ## Features
    - User registration and authentication
    - JWT token-based authentication (access and refresh tokens)
    - Password reset flow via email
    - Rate limiting protection
    - Comprehensive error handling

    ## Authentication
    Most endpoints require a valid JWT access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    Access tokens expire after 15 minutes. Use the refresh token endpoint to obtain new tokens.
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server
tags:
  - name: Authentication
    description: User authentication and token management endpoints
  - name: Password Reset
    description: Password reset flow endpoints
paths:
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account with email and password
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterUserSuccessResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticates a user and returns JWT tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserRequest'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccessResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh tokens
      description: Generates new access and refresh tokens using a valid refresh token
      operationId: refreshTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens successfully refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccessResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/request-password-reset:
    post:
      tags:
        - Password Reset
      summary: Request password reset
      description: Initiates password reset by sending an email with reset link
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPasswordResetRequest'
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageSuccessResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/reset-password:
    post:
      tags:
        - Password Reset
      summary: Reset password
      description: Completes password reset using a valid token and new password
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password successfully reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageSuccessResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    RegisterUserRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Email address for the new user account
          example: user@example.com
        password:
          type: string
          minLength: 4
          maxLength: 30
          description: Password for the new user account (minimum 4 characters, maximum 30 characters)
          example: SecurePassword123
    LoginUserRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Email address of the user
          example: user@example.com
        password:
          type: string
          minLength: 4
          maxLength: 30
          description: User password
          example: SecurePassword123
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Refresh token to generate a new token pair
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    RequestPasswordResetRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Email address to send password reset instructions
          example: user@example.com
    ResetPasswordRequest:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          description: Password reset token received via email
          example: 6b721a76-b1f8-4b0f-a07f-14f2d0792608
        newPassword:
          type: string
          minLength: 4
          maxLength: 30
          description: New password for the user account (minimum 4 characters, maximum 30 characters)
          example: NewPassword456
    RegisterUserResponse:
      type: object
      required:
        - id
        - email
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the created user
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          description: Email address of the user
          example: user@example.com
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the user was created
          example: '2024-03-27T10:30:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the user was last updated
          example: '2024-03-27T10:30:00Z'
    TokenPair:
      type: object
      required:
        - accessToken
        - refreshToken
        - tokenType
        - expiresIn
      properties:
        accessToken:
          type: string
          description: JWT access token for API authentication
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: JWT refresh token for obtaining new access tokens
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        tokenType:
          type: string
          description: Token type (always "Bearer")
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          description: Access token expiration time in seconds
          example: 900
    ErrorResponse:
      type: object
      required:
        - message
        - status
      properties:
        message:
          type: string
          description: Error message describing what went wrong
          example: Invalid credentials
        status:
          type: integer
          description: HTTP status code
          example: 400
    RegisterUserSuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          description: Indicates whether the request was successful
          example: true
        data:
          $ref: '#/components/schemas/RegisterUserResponse'
    LoginSuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          description: Indicates whether the request was successful
          example: true
        data:
          $ref: '#/components/schemas/TokenPair'
    MessageSuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          description: Indicates whether the request was successful
          example: true
        data:
          type: string
          description: Success message
          example: Password reset email sent